lambda_functions/v2/requirements/requirements.txt
#Update this date to trigger update of layers: 20220431
requests==2.24.0
pyjwt==2.4.0
boto3==1.22.2
requests-toolbelt==0.8.0
flask==2.1.1
flask-lambda==0.0.4
flask-lambda-python36==0.1.0
typing-extensions